+++ src/libstd/rt/stack.rs
