+++ src/libstd/unstable/mod.rs
@@ -19,7 +19,6 @@ pub mod finally;
 pub mod simd;
 pub mod sync;
 pub mod mutex;
-pub mod stack;
 
 /**
